+++ src/include/kernel/arch/common/mem/vmm_mapping.h
@@ -1,0 +1,11 @@
+#ifndef _VMM_MAPPING_H_
+#define _VMM_MAPPING_H_
+
+void *map_phys_to_virt_addr(void* physical_address, void* address, size_t flags, uint64_t *pml4_root);
+void identity_map_phys_address(void *pyhysical_address, size_t flags);
+void map_vaddress_range(void *virtual_address, size_t flags, size_t required_pages);
+
+void *map_vaddress(void *address, size_t flags);
+int unmap_vaddress(void *address);
+
+#endif
