+++ src/include/kernel/arch/x86_64/acpi.h
