+++ src/kernel/mem/kheap.c
@@ -4,6 +4,7 @@
 #include <vmm.h>
 #include <kernel.h>
 #include <logging.h>
+#include <vmm_mapping.h>
 #include <vmm_util.h>
 
 KHeapMemoryNode *kernel_heap_start;
@@ -20,7 +21,6 @@ void initialize_kheap(){
 
     kernel_heap_start = (KHeapMemoryNode *) ((uint64_t) kheap_vaddress);
     loglinef(Verbose, "(initialize_kheap) Start address using vmm_alloc: %x, and using end of vmm_space: %x", kheap_vaddress, kernel_heap_start);
-    loglinef(Verbose, "(initialize_kheap) PAGESIZE: 0x%x - val: %x", PAGE_SIZE_IN_BYTES, kernel_heap_start->size);
 
     #else
         #pragma message "(initialize_kheap) Using test specific initialization"
@@ -35,6 +35,7 @@ void initialize_kheap(){
     kernel_heap_current_pos->is_free = true;
     kernel_heap_current_pos->next = NULL;
     kernel_heap_current_pos->prev = NULL;
+    loglinef(Verbose, "(initialize_kheap) PAGESIZE: 0x%x - val: %x", PAGE_SIZE_IN_BYTES, kernel_heap_start->size);
 }
 
 size_t align(size_t size) {
@@ -93,9 +94,11 @@ void *kmalloc(size_t size) {
 }
 
 void expand_heap(size_t required_size) {
+    //  This function expand the heap in case more space is needed.
     loglinef(Verbose, "(expand_heap) called size: %d current_end: 0x%x - end_of_mapped_memory: 0x%x", required_size, kernel_heap_end, end_of_mapped_memory);
-    //size_t number_of_pages = required_size / KERNEL_PAGE_SIZE + 1;
+    // The first thing to do is compute how many page we need for this expansion
     size_t number_of_pages = get_number_of_pages_from_size(required_size);
+    // Then check where the heap ends
     uint64_t heap_end = compute_kheap_end();
         if ( heap_end > end_of_mapped_memory ) {
         //end_of_mapped memory marks the end of the memory mapped by the kernel loader.
